More data

+++ script.py
@@ -7,7 +7,7 @@ from timeit import Timer
 import time
 import socket
 
-N = Integer(1000)
+N = Integer(10000)
 maxmatrixsize = Integer(10)
 num_bound = Integer(100)
 den_bound = Integer(10)
